requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.0.2
certifi==2018.1.18
chardet==3.0.4
confusable-homoglyphs==3.0.0
Django==2.0.2
django-bootstrap-form==3.3
django-celery-beat==1.1.0
ephem==3.7.6.0
future==0.16.0
idna==2.6
kombu==4.1.0
mysqlclient==1.3.12
numpy==1.14.2
Pillow==5.0.0
pyaes==1.6.1
pyasn1==0.4.2
python-dateutil==2.6.1
python-telegram-bot==10.0.1
pytz==2018.3
redis==2.10.6
requests==2.18.4
rsa==3.4.2
scikit-learn==0.18.2
scipy==1.0.0
six==1.11.0
Telethon==0.17.3
Unidecode==1.0.22
urllib3==1.22
vine==1.1.4
